+++ parsers/type_parsing.lua
@@ -17,6 +17,8 @@ local keys = table_utils.keys
 local predicates = require("predicates")
 local isIdentifier = predicates.isIdentifier
 local isValueKeyword = predicates.isValueKeyword
+local isReservedName = predicates.isReservedName
+local isTupleFieldName = predicates.isTupleFieldName
 
 local error_reporting = require("error_reporting")
 local nullBadVal = error_reporting.nullBadVal
@@ -399,6 +401,14 @@ local function parse_type_record(badVal, xparsed, type_spec)
                         if isValueKeyword(field_name) then
                             utils.log(badVal, 'record', type_spec,
                                 "field name cannot be a keyword: "..field_name)
+                            fail = true
+                        elseif isReservedName(field_name) then
+                            utils.log(badVal, 'record', type_spec,
+                                "field name cannot be a reserved name: "..field_name)
+                            fail = true
+                        elseif isTupleFieldName(field_name) then
+                            utils.log(badVal, 'record', type_spec,
+                                "field name is reserved for tuples: "..field_name)
                             fail = true
                         else
                             fields_parsers[field_name] = value_val